requirements.txt
streamlit==1.21.0
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
requests==2.30.1
